feat: make any struct common to all api versions
