+++ docs/user/crd/index.html
@@ -5875,40 +5875,6 @@ Kubernetes admission/v1.Operation
 </tbody>
 </table>
 <hr />
-<h3 id="kyverno.io/v2.Any">Any
-</h3>
-<p>
-(<em>Appears on:</em>
-<a href="#kyverno.io/v2.Condition">Condition</a>)
-</p>
-<p>
-<p>Any can be any type.</p>
-</p>
-<table class="table table-striped">
-<thead class="thead-dark">
-<tr>
-<th>Field</th>
-<th>Description</th>
-</tr>
-</thead>
-<tbody>
-<tr>
-<td>
-<code>-</code><br/>
-<em>
-<a href="#kyverno.io/v2.Value">
-Value
-</a>
-</em>
-</td>
-<td>
-<em>(Optional)</em>
-<p>Value contains the value of the Any object.</p>
-</td>
-</tr>
-</tbody>
-</table>
-<hr />
 <h3 id="kyverno.io/v2.AnyAllConditions">AnyAllConditions
 </h3>
 <p>
@@ -6126,9 +6092,7 @@ Kubernetes meta/v1.Time
 <td>
 <code>key</code><br/>
 <em>
-<a href="#kyverno.io/v2.Any">
-Any
-</a>
+github.com/kyverno/kyverno/api/kyverno.Any
 </em>
 </td>
 <td>
@@ -6155,9 +6119,7 @@ DurationLessThanOrEquals, DurationLessThan</p>
 <td>
 <code>value</code><br/>
 <em>
-<a href="#kyverno.io/v2.Any">
-Any
-</a>
+github.com/kyverno/kyverno/api/kyverno.Any
 </em>
 </td>
 <td>
@@ -6615,14 +6577,6 @@ int
 </tbody>
 </table>
 <hr />
-<h3 id="kyverno.io/v2.Value">Value
-</h3>
-<p>
-(<em>Appears on:</em>
-<a href="#kyverno.io/v2.Any">Any</a>)
-</p>
-<p>
-</p>
 <h2 id="kyverno.io/v2alpha1">kyverno.io/v2alpha1</h2>
 <p>
 </p>
